Исправить проблемы с Telegram интерфейсом
- Исправлены callback buttons для кнопок 'Назад' и 'Главное меню'
- Добавлен правильный импорт CBT из tg_bot
- Переименован оригинальный плагин в backup для избежания конфликтов
- Рефакторированный плагин теперь основной (price_updater_plugin.py)

Исправления:
- ❌ Ошибка при клике 'Добавить лот' - исправлено
- ❌ Плашка 'PLUGIN_SETTINGS: 247153d9' при кнопках назад - исправлено
